Refactor calculating gross prices and shipping costs
Use the same logic for shipping rates as for line items: Upon creation
of the cost, apply the relevant included tax rates.

This resulted in some duplicate code, which is now extracted into a
concern and tested for not calling the database when it does not have
to.

Move calculating of gross amounts into a concern and test the concern
for calling `TaxRate.included_tax_amount_for` as little as possible.