Increase scale on `pre_tax_amount` columns
The `pre_tax_amount` column is used when calculating VATs. When the
precision is too low for this field, rounding errors occur.

For a line item with a price of 4.92 $ and a tax rate of 17%, the
calculation has to go like this:

```
4.92 / 1.17 = 4.205128205128205
line_item.save
-->
    pre_tax_amount: 4.205128205128205
    pre_tax_amount (rounded): 4.21
    included_tax_amount 71 ct.
    amount 4.92
```

Without this PR, whenever the line item is saved and reloaded, and then
the taxes are re-calculated, the calculation wrongly goes like this:

```
4.92 / 1.17 = 4.205128205128205
line_item.save
-->
    pre_tax_amount: 4.21
    included_tax_amount: 72 ct.
    amount: 4.92
```

That's an error: `4.21 + 72 != 2.92`.

This should not affect any display_* methods, as those use
`Spree::Money`, which takes care of good rounding.

Fixes #6472