Link to RMAs instead of "Returns"
On the edit order page, there's a link to the order's return authorizations.
This link is translated as "returns", which is confusing (the German translation
for example now has two links both saying "customer returns").

The key is only ever used in this instance, so when we change it to RMAs,
we can get rid of one translation key.