Add credit method to Check payment method
When using the `Check` payment method, you can not create a reimbursement
as the `ReimbursementPerfomer` always creates a `Spree::Refund`, which calls
the `#credit` method on the payment method originally used.

This fix adds a `credit` method on the `check` payment method.

It also refactors the simulated successful billing response into a private method.

Fixes #6529