Destroy shipments when verifying available rates
Prefer correctness over premature optimization.

With the delete_all there is not a cascade delete of related objects including the inventory unit.
When building an order on the backend this causes orphaned inventory unit records which
manifest themselves confusingly in the RMA area as you have rows for each line item.

This was noted as a comment[1] on the original commit and a PR was requested. This is that PR.

1. https://github.com/spree/spree/commit/aadb5ae66eb1f51b563c5d998a9d8a53c731c241#commitcomment-9487805

Fixes #6592