Namespace Order Events
In German, to 'cancel' an order has to be a specific word ("stornieren") that is different from e.g.
canceling a UI dialog ("abbrechen"). Since "cancel" is used in the `event_links` helper for
the `Cancel order` button, I get a (for a German) slightly confusing thing that hints at me being able
to somehow "close" an order.

This commit introduces namespacing just for this helper. I will also provide a PR to provide
the necessary translations to the German translation in `spree_i18n`.

Since other Spree extension might be using this helper and only provide non-namespaced event translations,
I call `Spree.t` with `default: Spree.t(event)`.

Fixes #6615