More fully reset XHR populated fields
If you switch customers it changes some values (such as e-mail) but the billing address field
is only changed if the new user has a billing field. This will clear out the existing billing fields
if the new user doesn't have a billing address.

Also the shipping address is never updated. This change clears out the shipping address of
possibly old info and reverts to the assumption that we want to use the billing address for
shipping.