Optimize OrderUpdater#update_adjustment_total
In our specific use case this halved the execution time of this method
under profiling. The trick here is to replace a number of activerecord
sum calls with a pluck and reduce.

This optimisation is best for small orders because doing a constant 2
activerecord calls is cheaper than doing 8 sum calls. This will get more
expensive as the basket size grows but unless people have thousands of
items it's unlikely to get pathological.