Allow more options than tax zone for price calculation
Our client needs to display different prices to different customers. The MOSS
refactoring has enabled different prices for different tax zones, which is great,
but they also need to display prices excluding VAT for *some* business customers.

The logic for that is (while in accordance with EU tax law) rather custom. This PR
is supposed only to enable passing more than just the tax zone as options for price calculation.

In order to accomplish this, the `current_tax_zone` helper is replaced with a `current_price_options`
helper which returns a Hash. All the values of that Hash are used to calculate product cache keys.
I took care to order those so that we don't end up with different caches for the same content.

* Remove current_tax_zone helper and replace with current_price_options
* Adapt product cache key to price options hash
* Adapt base helper to call #price_including_vat_for with price_options hash
* Adapt price model for price_options hash
* Adapt VAT price calculation module
* Use new #included_tax_amount signature
* Adapt tax_rate.rb
* Adapt line_item.rb
* Add tests for edge case price options
* Adapt StockEstimator