Fix update! returning nil when !source.present.
Bug: Calling update! on an adjustment with no source always returns nil.

Ruby scopes variables by block scope, and onditionals do not open a new scope.
Essentially, what happens in this function, is that an `amount` var is
immediately defined (at the top of the function) and set to nil.

This then completely takes precedence over the `amount` instance method,
returning nil instead of the actual amount.

The fix is to use `self.amount` when refering to the method, although a much
better practice would be to not use vars that match real variables.

We have done no testing on 3.0, where the method has been refactored. Chances
are that it's suffering from the same bug.

POC code, displaying the bug:

  class A
    def amount
      1
    end
    def amount=
      "whatever"
    end
    def update!
      return amount if false
      if false
        amount = 2
      end
      amount
    end
  end

  A.new.update! # => nil

Also note that even if source was present, `amount=` never gets called because
`amount = 2` creates a local variable, while `self.amount = 2` would call the
function.