Changes the default value for Spree::Adjustment.included to false.
The database already did this, but Spree 3.0 moved the creation of Promotion Adjustments into AdjustmentSource.create_adjustment. The nil value would be saved to the DB, ignoring the default.

This caused orders with line item promotions to fail when attempting to use PayPal.

This change should be applied to master and 3-0-stable.