Display output of state to match the form
It seems if frontend/app/views/spree/address/_form.html.erb hides the state completely when this config flag
is enabled we should also hide it here. This was causing confusion for some UK users were the "state" was
not desired but due to automatic data population from another system it was populating the state with the
"town". So users were seeing info they did not enter and had not be requested.

This commit updates the templates to match for consistency.