Make refreshing rates safer for the backend.
If a shipping option is only available for the backend you can easily
loose that selection. Visiting the cart page on an incomplete order
results it is being lost since the rate refresh uses the default of
frontend only.

Also if you manaully trigger an shipping update it will be lost since
there is not an option to specify if backend options should be included.
This commit allows that option to be specified.