Allow payment sources that are not credit cards
We have a number of reusable payment sources that are not credit
cards. However, the `payment_sources` method on the `Spree.user_class`
model is a simple `has_many` that can only get Credit cards.

This PR adds a join table between the user and the different payment source
models to accomodate for the ensuing Polymorphism. Common code for all
payment sources is abstracted into a `Spree::PaymentSource` base class,
and the common tests are extracted into `shared_examples` for all Payment sources.

Consequently, also the code for re-using existing payment sources in
`order/checkout.rb` had to be adapted.

This is a rather large PR. Please have a careful look at it; I'm happy to integrate
comments and criticism.