Refactoring the way we try to build packages
We were iterating over each one of the active stock locations to build packages and skipping those without the variant in stock.

This was triggering a lot of unnecessary queries to database and this commit address this issue.

Now we try find just the stock locations that contain the requested variants in stock and build packages for each one of them.