Remove text_for_button_link helper
This used to give a button and supported adding icons to the text, but
as of 14c3b622daca7bbf92076ea2a70fd34bea9233c4 it does nothing (other
than use raw which could be dangerous).

This is never used in a dangerous way in the project itself, but could
potentially be dangerous when used in custom admin components.

Conflicts:
	backend/app/helpers/spree/admin/navigation_helper.rb