Simplify javascript detection in address form
A user could input a specially crafted string here to inject HTML on the
page.

In normal circumstances this should be difficult to exploit because of
our CSRF protection on form submissions.

This could be exploited on stores which have a "masquerade" function
allowing admins to browse the store and checkout as a user.

Conflicts:
	frontend/app/assets/javascripts/spree/frontend/checkout/address.js.coffee