Admin Panel: unified order completion DateTime format - not with time
Previously store owners to check the time of placing an order would
have to visit the order page which wasn't a good user experience
choice. Now it displays date and time in condensed format so it also
doesn't cost a lot of screen space. For this, there is a new helper
method called order_time used everywhere in the Admin Panel for
displaying order completion date & time.