Do not apply user-coupled promotions to all carts
The code in the Cart promotion handler will first try if a promotion is
eligible for the line item. However, Promotion#eligible? is implemented
in such a way that it will consider itself eligible for a line item
if it is only applicable to orders.

Therefore, a user-related promotion without a promotion code, which
is not applicable to line items, will be considered eligible for all
orders. If there are many such promotions, this means each order will be
burdened by a large number of inactive promotion adjustments.

This change fixes this problem by checking for eligibility with both
line item and order.