Fix exceptions on checkout workflow when going direct to link
As described by bug 7176, when a user goes directly to a checkout link
without going to the cart and clicking through the user interface, they
can get an exception. Reproduction instructions are available on the bug
comment at
https://github.com/spree/spree/issues/7176#issuecomment-196119351

This patch fixes the bug by moving the check for checking whether a
state is valid to BEFORE setting the order state from the user
parameters. In addition, the unknown_state? method is changed to test
the order state regardless of whether the user has supplied a state in
the parameters.

Fixes #7176