Replace cache_key_for_taxons with cache
Rails 5.0 supports generating a cache key from a scope, using a similar
technique to what cache_key_for_taxons was using. This commit
deprecates cache_key_for_taxons favour of using cache with a scope
directly.

Passing a scope or collection to `cache` will generate a hash based on
the scope and MAX(updated_at) and COUNT(*). This is better than the
existing helper because it takes the full scope into account and finds
the max_updated_at and count values in a single query.