Provide correct HTTP code with failed coupon code
When a coupon code fails it is responding with a HTTP 200 instead of a 422. 422 is used
elsewhere in this file (for example on the could not transition error). Also the `apply_coupon_code`
action in the Order api controller returns 422. This just makes this consistent.

I noticed above this that the insufficient quantity error has the same problem so went ahead and hit
that one also.