Fixes #8502
This commit fixes the searching of the taxons when the request searching
within a `taxonnomy`.

The code is re-arranged a bit to DRY the ransack call.

There is a `skipped` test which is the case of searching for nested
taxons. I am leaving that open for discussion in this PR.