Add presence validation to display_on of ShippingMethod
* Why was this change necessary?
ShippingMethod will be picked up for calculating pricing for an order.
So if it doesn't have value on display_on, it leads us to get some
misleading error message (https://github.com/spree/spree/issues/9211).
So this pr, we attempt to fix this error by adding a presence validation
for it. To void conflicting with this PR as well
https://github.com/spree/spree/pull/8430

* How does it address the problem?
By adding a presence validation for display_on, we will make sure that,
the old shipping method won't be affected, it will still shown as None
value in DisplayOn.

But this PR will force the admin, they have to pick up a suitable
display on method for shipping method instead of leave it blank