[SDP-385] Fixes products carousel caching problem
The problem occurred on related products carousel on PDP. It was
noticed that removing a product from related products list for
product resulted in showing a product that was not present on
the list of related products. This has happened after related
products were added using a script and multiple products in
this list had the same `updated_at` dates.