Change `includes` to `preload` in Spree::Admin::ImagesController
Since `includes` will try to pick the best strategy for joining the
associated models, it will first try to use `eager_load` and it will fail
because `:viewable` is a polymorphic association. Therefore, we have to tell
rails to use preload as a strategy which can handle a polymorphic association.