Fix flatpickr intermittent failing tests. (#11223)
Use exact text when selecting day to avoid ambiguous match on single digit dates.
Add small pause between changing flatpickr values that adjust the DOM in the calendar, to ensure JavaScript to updates the DOM before next section is made.
Select month by value=, not by nth-child (**1)
Use find not xpath.
REFACTOR: Rather than passing in a string and chopping it up, now pass an options hash specifying the day: 4, month: 3, year: 2018. This gives greater clarity regardless if you use U.S. or European date standards as to what the values are actually being sent.
REFACTOR: Reduce 2 methods into one, instead of calling fill_in_date_time_picker( ... ), now just pass in the hour: 11, minute: 45 along with the day: 4, month: 3, year: 2018 and fill_in_date_picker will recognise the presence of the additional time values and set the time.
(**1) If the From and To dates are close together, only the available months are populated in the options, so if nth-child(9) (September) is looked up, but the months start from September, then September would be the nth-child(1) and nth-child(9) would not exist, as per the image attached.

Added back in previously broken test for filter using date pickers, since refactor this is now testable.