Check if variant was in stock before update
The variant.out_of_stock event should be emitted only when the variant
was in stock before the update and now it is not. This is important in
case the variant is out of stock and after an update it still persists 0
as stock. In this case the event was already emitted and should not be
emitted again.