SD-1511 System should emit `variant.out_of_stock` if Variant is sold out (#11464)
* Emit variant.out_of_stock webhook event

* Remove unnecessary time freezing in spec

* Check if variant was in stock before update

The variant.out_of_stock event should be emitted only when the variant
was in stock before the update and now it is not. This is important in
case the variant is out of stock and after an update it still persists 0
as stock. In this case the event was already emitted and should not be
emitted again.

* Do not use #reload in decorator

* Check if variant in stock before update

* Add spec for variant out of stock before and after update