Disable webhooks for block
Defined a Spree::Webhooks.disable_webhooks method that takes in a
block of code. It disables the webhook events using the DISABLE_SPREE_WEBHOOKS
environment variable. After that it runs the block and then sets the
environment variable to its original value. If an error is raised
inside the given block, we do not handle it and let the caller do the
handling. In any given case (whether the block is executed fully or
it is interrupted), we set the environment variable to its original
value