Delegate OrderPromotion.public_metadata to Order.public_metadata
Issue arises when trying to serialize the Order (Cart). Namely it tries to
serialize OrderPromotion when promotions is included, and when doing that
it can not find the .public_metadata, so we need to delegate either to order
or promotion. Maybe we should have a OrderPromotion serializer like for Platform
API.