Fix created_at null error running CreateSpreeProductsStores migration
In #11363 there was an attempt to fix this issue. At that time the fix
only covered the case in which the `created_at` column already existed,
but it did not add a default value for timestamps in the `create_table`
block so that it is taken into account when the table does not exist at
all. This patch provides a default value also in that case.