Fix address not being saved in v2 controller
If we use the current checkout controller v2, a new address will not be automatically saved on the user. Although on line #33 and #48 the user.bill/ship_address is updated, this is not persisted to the databse. Instead of `user.bill_address` we'd prefer to have it saved to `user.addresses` as well. 

I think the current implementation of `update_or_create_address` also could inadvertedly return an address from another user. There could be a (hypothetical) race condition, where another user edits this address while the order is not yet complete.