Separate model for Cms Section images
New CmsSectionImage models for each image were created, because when we have
just one new model and since it is polymorphic, we have the same viewable_id for each one.
When we now create a CmsSectionImage a webhook is emitted (since it is an Asset) and it tries
to serialize it with a name corresponding to the class. That is why I have added the change to
specify the serializer. Alternative would be to climb up the descendants and try to find a serializer.
I find this approach simpler (specify in a class method). Also when creating that new Asset, the viewable
relationship is included, hence it tries to serialize HeroImages. That is why I have created those new
serializers that just inherit from CmsSectionSerializer.