Show more specific error when adding item to cart fails
When you try to add an item to the cart that is cannot be supplied
you get an error that the Quantity is not available. There is the case
when you have the available quantity, but the variant is not .available?
which happens when the product has status different from Active (Draft or
Archived). For this case we also consider that the item cannot be supplied
but we should specify that it is because of the status and not the quantity.